+++ Classes/TemperatureUnitTableViewController.m
@@ -156,8 +156,8 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
    
-	int newRow = [indexPath row];
-	int oldRow = [lastIndexPath row];
+	NSInteger newRow = [indexPath row];
+	NSInteger oldRow = [lastIndexPath row];
 	UserModel *userModel = [UserModel userModel];
 	
 	if (newRow != oldRow)
